refactor: outsource Drawer as its own component
