+++ pages/inventors/index.js
@@ -5,7 +5,7 @@ import Layout from '@/layouts/main';
 import { Icon } from '@iconify/react';
 import Image from 'next/image';
 import Button from '@/components/Button';
-import ViewProfile from '@/components/view-profile';
+import ViewProfile from '@/components/ViewProfile';
 
 import { inventorsLeads } from '@/utils/leads';
 import { shortenEmail, shortenPhone } from '@/utils/helpers';
